Simplified test task.
