+++ lib/tasks/client/build-client/index.js
@@ -7,7 +7,7 @@ const buildClient = function () {
   gulp.task('_build-client', function (done) {
     runSequence(
       'analyze',
-      'test-units',
+      'test',
       'clean-client',
       [ 'build-scripts', 'build-themes', 'build-html', 'copy-static' ],
       done);
