+++ lib/tasks/server/build-server/index.js
@@ -7,7 +7,7 @@ const buildServer = function () {
   gulp.task('_build-server', function (done) {
     runSequence(
       'analyze',
-      'test-units',
+      'test',
       'unused-dependencies',
       'outdated',
       done);
