+++ lib/tasks/server/watch-server/index.js
@@ -8,7 +8,7 @@ const watchServer = function (roboter) {
     roboter.isWatching = true;
 
     runSequence(
-      'watch-test-units',
+      'watch-test',
       done);
   });
 };
