+++ lib/tasks/universal/test/index.js
@@ -1,14 +1,33 @@
 'use strict';
 
-const gulp = require('gulp'),
+const _ = require('lodash'),
+    gulp = require('gulp'),
     runSequence = require('run-sequence');
 
-const test = function () {
+const defaultConfiguration = {
+  watch: [ './**/*.js', '!./node_modules/**/*.js' ]
+};
+
+const test = function (roboter, userConfiguration) {
+  const configuration = _.assign({}, defaultConfiguration, userConfiguration);
+
   gulp.task('_test', function (done) {
     runSequence(
       'test-units',
       'test-integration',
       done);
+  });
+
+  gulp.task('_watch-test-continuously', function (done) {
+    runSequence(
+      'test-units-continuously',
+      'test-integration-continuously',
+      done);
+  });
+
+  gulp.task('_watch-test', function () {
+    roboter.isWatching = true;
+    gulp.watch(configuration.watch, [ 'watch-test-continuously' ]);
   });
 };
